% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralplottingtool.R
\name{spectralplottingtool}
\alias{spectralplottingtool}
\title{Plotting function to output PNG or an R object of the spectrum.}
\usage{
spectralplottingtool(
  flowfile,
  theme = "viridis",
  save = FALSE,
  bins = 512,
  normalize = FALSE,
  params = NULL
)
}
\arguments{
\item{flowfile}{A flowSet or flowFrame}

\item{theme}{Choose a theme: 'viridis' (default), 'bigfoot', or 'aurora'}

\item{save}{FALSE: in console (default). TRUE : as png file in working directory}

\item{bins}{Choose the granularity of the data.  Between 200 and 1000 works well for most data.}

\item{normalize}{Normalize the data to the max median signal (only for clean signals).  TRUE or FALSE (default)}

\item{params}{Default NULL. Specify parameters (in order) to plot. Only used if data are not from an Aurora, ID7000, or BigFoot instrument.}
}
\value{
Images of full spectrum
}
\description{
Can be wrapped in the data loading function, or directly with a
flowFrame or inside fsApply() for a flowSet.
This function tidies the fcs data from a flowFrame or flowSet.
Choose a theme: 'viridis', 'bigfoot', or 'aurora'.  Viridis is
colourblind accessible.
Choose to save to the working directory or display in current session.
Choose the number of bins (granularity) of the data.
You can normlize to the max median signal for "clean" files.
If you misspell an option it will give an error.
}
