% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralplot.R
\name{spectralplot}
\alias{spectralplot}
\title{Load data, choose a style, and choose output method}
\usage{
spectralplot(
  flowfile,
  theme = "viridis",
  save = FALSE,
  bins = 512,
  normalize = FALSE,
  params = NULL
)
}
\arguments{
\item{flowfile}{A flowSet or flowFrame}

\item{theme}{Choose a theme: 'viridis' (default), 'bigfoot', or 'aurora'}

\item{save}{FALSE: in console (default). TRUE : as png file in working directory}

\item{bins}{Choose the granularity of the data.}

\item{normalize}{Normalize the data to the max median signal (only for clean signals).  TRUE or FALSE (default)}

\item{params}{Default NULL. Specify parameters (in order) to plot. Only used if data are not from an Aurora, ID7000, or BigFoot instrument.}
}
\value{
Images of full spectrum
}
\description{
This function loads a flowFrame of flowSet and invokes the plotting function.
Choose a theme: 'viridis', 'bigfoot', or 'aurora'.  Viridis is
colourblind accessible.
Choose to save to the working directory or display in current session.
Choose the number of bins (granularity) of the data.
If you misspell an option it will give an error.
}
