% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralplot.R
\name{spectralplot}
\alias{spectralplot}
\title{Load data, choose a style, and choose output method}
\usage{
spectralplot(flowfile, theme = "viridis", save = FALSE)
}
\arguments{
\item{flowfile}{A flowSet or flowFrame}

\item{theme}{Choose a theme: 'viridis' (default), 'bigfoot', or 'aurora'}

\item{save}{FALSE: in console (default). TRUE : as png file in working directory}
}
\value{
Images of full spectrum
}
\description{
This function loads a flowFrame of flowSet and invokes the plotting function.
Choose a theme: 'viridis', 'bigfoot', or 'aurora'.  Viridis is
colourblind accessible.
Choose to save to the working directory or display in current session.
If you misspell an option it will give an error.
}
